requirements.txt
streamlit==1.46.0
websocket-client==1.8.0